DeclerationFees

+++ GAZT/GAZT/Models/EDeclerationsModel/FeesCalculators/FeesCalculatorResponse.cs
@@ -8,13 +8,13 @@ namespace EGAZT.Models.EDeclerationsModel.FeesCalculators
     {
         public string tobaccoExciseTaxEquation { get; set; }
         public string tobaccoCustomsTaxEquation { get; set; }
-        public int? customsPercentage { get; set; }
-        public int? productFinalPrice { get; set; }
+        public double? customsPercentage { get; set; }
+        public double? productFinalPrice { get; set; }
         public double totalPayment { get; set; }
         public double vat { get; set; }
-        public int? extraFees { get; set; }
+        public double? extraFees { get; set; }
         public double excise { get; set; }
-        public int? totalDuty { get; set; }
+        public double? totalDuty { get; set; }
     }
 }
